+++ mac/starsky/Services/FileWatcherServiceProtocol.swift
@@ -1,6 +1,7 @@
 import Foundation
 
-protocol FileWatcherServiceProtocol: AnyObject {
+// @unchecked Sendable: stop() is called from a background queue during termination by design.
+protocol FileWatcherServiceProtocol: AnyObject, Sendable {
     func start()
     func stop()
 }